+++ packages/apollo-collaboration-server/src/export/export.controller.spec.ts
@@ -1,0 +1,19 @@
+import { Test, TestingModule } from '@nestjs/testing'
+
+import { ExportController } from './export.controller'
+
+describe('ExportController', () => {
+  let controller: ExportController
+
+  beforeEach(async () => {
+    const module: TestingModule = await Test.createTestingModule({
+      controllers: [ExportController],
+    }).compile()
+
+    controller = module.get<ExportController>(ExportController)
+  })
+
+  it('should be defined', () => {
+    expect(controller).toBeDefined()
+  })
+})
