+++ packages/apollo-collaboration-server/src/export/export.module.ts
@@ -1,0 +1,23 @@
+import { Module } from '@nestjs/common'
+import { MongooseModule } from '@nestjs/mongoose'
+import { Export, ExportSchema } from 'apollo-schemas'
+
+import { AssembliesModule } from '../assemblies/assemblies.module'
+import { FeaturesModule } from '../features/features.module'
+import { RefSeqChunksModule } from '../refSeqChunks/refSeqChunks.module'
+import { RefSeqsModule } from '../refSeqs/refSeqs.module'
+import { ExportController } from './export.controller'
+import { ExportService } from './export.service'
+
+@Module({
+  imports: [
+    AssembliesModule,
+    FeaturesModule,
+    MongooseModule.forFeature([{ name: Export.name, schema: ExportSchema }]),
+    RefSeqsModule,
+    RefSeqChunksModule,
+  ],
+  providers: [ExportService],
+  controllers: [ExportController],
+})
+export class ExportModule {}
