+++ packages/apollo-collaboration-server/src/export/export.service.spec.ts
@@ -1,0 +1,19 @@
+import { Test, TestingModule } from '@nestjs/testing'
+
+import { ExportService } from './export.service'
+
+describe('ExportService', () => {
+  let service: ExportService
+
+  beforeEach(async () => {
+    const module: TestingModule = await Test.createTestingModule({
+      providers: [ExportService],
+    }).compile()
+
+    service = module.get<ExportService>(ExportService)
+  })
+
+  it('should be defined', () => {
+    expect(service).toBeDefined()
+  })
+})
