+++ packages/apollo-collaboration-server/src/features/features.module.ts
@@ -1,9 +1,8 @@
 import { Module, forwardRef } from '@nestjs/common'
 import { MongooseModule, getConnectionToken } from '@nestjs/mongoose'
-import { Export, ExportSchema, Feature, FeatureSchema } from 'apollo-schemas'
+import { Feature, FeatureSchema } from 'apollo-schemas'
 import idValidator from 'mongoose-id-validator'
 
-import { AssembliesModule } from '../assemblies/assemblies.module'
 import { OperationsModule } from '../operations/operations.module'
 import { RefSeqChunksModule } from '../refSeqChunks/refSeqChunks.module'
 import { RefSeqsModule } from '../refSeqs/refSeqs.module'
@@ -14,8 +13,6 @@ import { FeaturesService } from './features.service'
   controllers: [FeaturesController],
   providers: [FeaturesService],
   imports: [
-    // AssembliesModule,
-    // forwardRef(() => AssembliesModule), //We need this only to get assembly friendly name for exported GFF3 file
     forwardRef(() => OperationsModule),
     RefSeqChunksModule,
     RefSeqsModule,
@@ -28,7 +25,6 @@ import { FeaturesService } from './features.service'
         },
         inject: [getConnectionToken()],
       },
-      { name: Export.name, useFactory: () => ExportSchema },
     ]),
   ],
   exports: [MongooseModule],
